address remaining PR feedback
