+++ 3rdparty/quickjs/quickjs.c
@@ -34691,25 +34691,7 @@ int JS_SetModuleExportList(JSContext *ctx, JSModuleDef *m,
     return 0;
 }
 
-// https://www.freelists.org/post/quickjs-devel/Patch-Add-some-informations-to-get-more-informations-from-compiled-modules
-int JS_GetModuleExportEntriesCount(JSModuleDef *m)
-{
-    return m->export_entries_count;
-}
-
-JSValue JS_GetModuleExportEntry(JSContext *ctx, JSModuleDef *m, int idx)
-{
-    if (idx >= m->export_entries_count || idx < 0)
-        return JS_UNDEFINED;
-    return JS_DupValue(ctx, m->export_entries[idx].u.local.var_ref->value);
-}
-
-JSAtom JS_GetModuleExportEntryName(JSContext *ctx, JSModuleDef *m, int idx)
-{
-    if (idx >= m->export_entries_count || idx < 0)
-        return JS_ATOM_NULL;
-    return JS_DupAtom(ctx, m->export_entries[idx].export_name);
-}
+#include "quickjs-exports.c"
 
 /* Note: 'func_obj' is not necessarily a constructor */
 static void JS_SetConstructor2(JSContext *ctx,
