+++ 3rdparty/quickjs/quickjs.h
@@ -999,10 +999,6 @@ int JS_SetModuleExport(JSContext *ctx, JSModuleDef *m, const char *export_name,
 int JS_SetModuleExportList(JSContext *ctx, JSModuleDef *m,
                            const JSCFunctionListEntry *tab, int len);
 
-// https://www.freelists.org/post/quickjs-devel/Patch-Add-some-informations-to-get-more-informations-from-compiled-modules
-int JS_GetModuleExportEntriesCount(JSModuleDef *m);
-JSValue JS_GetModuleExportEntry(JSContext *ctx, JSModuleDef *m, int idx);
-JSAtom JS_GetModuleExportEntryName(JSContext *ctx, JSModuleDef *m, int idx);
 
 #undef js_unlikely
 #undef js_force_inline
